Dev: removed Controls class. Merged those methods into GrapherView. Moved AddEntityButton to its own file
