+++ grapher/controls/AddEntityButton.tsx
@@ -1,0 +1,84 @@
+import * as React from "react"
+import { observer } from "mobx-react"
+
+type HorizontalAlign = "left" | "right"
+type VerticalAlign = "top" | "middle" | "bottom"
+
+@observer
+export class AddEntityButton extends React.Component<{
+    x: number
+    y: number
+    align: HorizontalAlign
+    verticalAlign: VerticalAlign
+    height: number
+    label: string
+    onClick: () => void
+}> {
+    static defaultProps = {
+        align: "left",
+        verticalAlign: "bottom",
+        height: 21,
+        label: "Add country",
+    }
+
+    static calcPaddingTop(
+        y: number,
+        verticalAlign: VerticalAlign,
+        height: number
+    ): number {
+        const realY =
+            verticalAlign === "bottom"
+                ? y - height
+                : verticalAlign === "middle"
+                ? y - height / 2
+                : y
+        return Math.max(0, -realY)
+    }
+
+    render() {
+        const {
+            x,
+            y,
+            align,
+            verticalAlign,
+            height,
+            label,
+            onClick,
+        } = this.props
+
+        const buttonStyle: React.CSSProperties = {
+            position: "absolute",
+            lineHeight: `${height}px`,
+        }
+
+        if (verticalAlign === "top") {
+            buttonStyle.top = `${y}px`
+        } else if (verticalAlign === "bottom") {
+            buttonStyle.top = `${y - height}px`
+        } else {
+            buttonStyle.top = `${y - height / 2}px`
+        }
+
+        if (align === "left") {
+            buttonStyle.left = `${x}px`
+        } else if (align === "right") {
+            buttonStyle.right = `${-x}px`
+        }
+
+        return (
+            <button
+                className="addDataButton clickable"
+                onClick={onClick}
+                data-track-note="chart-add-entity"
+                style={buttonStyle}
+            >
+                <span className="icon">
+                    <svg width={16} height={16}>
+                        <path d="M3,8 h10 m-5,-5 v10" />
+                    </svg>
+                </span>
+                <span className="label">{label}</span>
+            </button>
+        )
+    }
+}
